Cleared the current set of build errors (no guarantees the world can gen).
Added new class of alert-based events.

+++ Capabilities/Stealth/Senses/Sixth.cs
@@ -24,7 +24,7 @@ namespace StealthSystemPrototype.Senses
         {
         }
 
-        public override double GetIntensity()
+        public override int GetIntensity()
             => base.GetIntensity();
 
         protected override ISense Copy()
